[Refactor] Delete unused routes
