--- src/desktop/apps/artwork/components/buyers_premium/index.styl
@@ -1,10 +1,0 @@
-@require '../../../../../../components/buyers_premium'
-
-.artwork-auction-buyers-premium-modal
-  //
-
-  .modalize-body
-    padding (gutter * 3)
-
-    h1
-      sans('l-headline')
