+++ src/desktop/apps/auction_support/index.coffee
@@ -1,6 +1,5 @@
 express = require 'express'
 routes = require './routes'
-{ skipIfClientSideRoutingEnabled } = require("../../components/split_test/skipIfClientSideRoutingEnabled")
 
 app = module.exports = express()
 app.set 'views', __dirname + '/templates'
