--- src/desktop/apps/search2/client.tsx
@@ -1,26 +1,0 @@
-import { buildClientApp } from "v2/Artsy/Router/client"
-import { routes } from "v2/Apps/Search/routes"
-import { data as sd } from "sharify"
-import React from "react"
-import ReactDOM from "react-dom"
-
-const mediator = require("desktop/lib/mediator.coffee")
-
-buildClientApp({
-  routes,
-  context: {
-    user: sd.CURRENT_USER,
-    mediator,
-  },
-})
-  .then(({ ClientApp }) => {
-    ReactDOM.hydrate(<ClientApp />, document.getElementById("react-root"))
-    document.getElementById("loading-container").remove()
-  })
-  .catch(error => {
-    console.error(error)
-  })
-
-if (module.hot) {
-  module.hot.accept()
-}
