--- src/desktop/apps/search2/templates/index.jade
@@ -1,9 +1,0 @@
-extends ../../../components/main_layout/templates/redesign
-
-append locals
-  - assetPackage = 'search2'
-
-block body
-  #react-root
-  #loading-container
-    != loadingComponent
