--- src/desktop/apps/viewing-room/client.tsx
@@ -1,28 +1,0 @@
-import { buildClientApp } from "v2/Artsy/Router/client"
-import { data as sd } from "sharify"
-import { routes as viewingRoomRoutes } from "v2/Apps/ViewingRoom/routes"
-import React from "react"
-import ReactDOM from "react-dom"
-import { loadableReady } from "@loadable/component"
-
-const mediator = require("desktop/lib/mediator.coffee")
-
-buildClientApp({
-  routes: viewingRoomRoutes,
-  context: {
-    user: sd.CURRENT_USER,
-    mediator,
-  },
-})
-  .then(({ ClientApp }) => {
-    loadableReady(() => {
-      ReactDOM.hydrate(<ClientApp />, document.getElementById("react-root"))
-    })
-  })
-  .catch(error => {
-    console.error(error)
-  })
-
-if (module.hot) {
-  module.hot.accept()
-}
