--- src/desktop/components/split_test/skipIfClientSideRoutingEnabled.ts
@@ -1,3 +1,0 @@
-export const skipIfClientSideRoutingEnabled = (_req, res, next) => {
-  return next("route")
-}
